backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
python-multipart==0.0.9
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
groq==0.9.0
httpx==0.26.0
